Changed: Make ImageStyleUI independent of ImageStyleEditing.
