+++ tests/imagestyle/imagestyleediting.js
@@ -7,22 +7,16 @@ import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtest
 import ImageStyleEditing from '../../src/imagestyle/imagestyleediting';
 import ImageEditing from '../../src/image/imageediting';
 import ImageStyleCommand from '../../src/imagestyle/imagestylecommand';
-import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import log from '@ckeditor/ckeditor5-utils/src/log';
+
 import { getData as getModelData, setData as setModelData } from '@ckeditor/ckeditor5-engine/src/dev-utils/model';
 import { getData as getViewData } from '@ckeditor/ckeditor5-engine/src/dev-utils/view';
-
-import fullWidthIcon from '@ckeditor/ckeditor5-core/theme/icons/object-full-width.svg';
-import leftIcon from '@ckeditor/ckeditor5-core/theme/icons/object-left.svg';
-import centerIcon from '@ckeditor/ckeditor5-core/theme/icons/object-center.svg';
-import rightIcon from '@ckeditor/ckeditor5-core/theme/icons/object-right.svg';
 
 import testUtils from '@ckeditor/ckeditor5-core/tests/_utils/utils';
 
 testUtils.createSinonSandbox();
 
 describe( 'ImageStyleEditing', () => {
-	let editor, plugin, model, document, viewDocument;
+	let editor, model, document, viewDocument;
 
 	afterEach( () => {
 		editor.destroy();
@@ -263,7 +257,7 @@ describe( 'ImageStyleEditing', () => {
 		} );
 	} );
 
-	describe( 'imageStyles()', () => {
+	describe( 'config', () => {
 		it( 'should fall back to defaults when no image.styles', () => {
 			return VirtualTestEditor
 				.create( {
@@ -308,188 +302,6 @@ describe( 'ImageStyleEditing', () => {
 
 					expect( newEditor.config.get( 'image.styles' ) ).to.deep.equal( [ { name: 'imageStyleSide' } ] );
 				} );
-		} );
-
-		it( 'should cache the styles', () => {
-			return VirtualTestEditor
-				.create( {
-					plugins: [ ImageStyleEditing ]
-				} )
-				.then( newEditor => {
-					editor = newEditor;
-					plugin = editor.plugins.get( ImageStyleEditing );
-
-					expect( plugin.imageStyles ).to.equal( plugin.imageStyles );
-				} );
-		} );
-
-		describe( 'object format', () => {
-			beforeEach( () => {
-				class TranslationMock extends Plugin {
-					init() {
-						sinon.stub( this.editor, 't' ).returns( 'Default translation' );
-					}
-				}
-
-				return VirtualTestEditor
-					.create( {
-						plugins: [ TranslationMock, ImageStyleEditing ],
-						image: {
-							styles: [
-								// Custom user styles.
-								{ name: 'foo', title: 'foo', icon: 'custom', isDefault: true, className: 'foo-class' },
-								{ name: 'bar', title: 'bar', icon: 'right', className: 'bar-class' },
-								{ name: 'baz', title: 'Side image', icon: 'custom', className: 'baz-class' },
-
-								// Customized default styles.
-								{ name: 'imageStyleFull', icon: 'left', title: 'Custom title' }
-							]
-						}
-					} )
-					.then( newEditor => {
-						editor = newEditor;
-						plugin = editor.plugins.get( ImageStyleEditing );
-					} );
-			} );
-
-			it( 'should pass through if #name not found in default styles', () => {
-				expect( plugin.imageStyles[ 0 ] ).to.deep.equal( {
-					name: 'foo',
-					title: 'foo',
-					icon: 'custom',
-					isDefault: true,
-					className: 'foo-class'
-				} );
-			} );
-
-			it( 'should use one of default icons if #icon matches', () => {
-				expect( plugin.imageStyles[ 1 ].icon ).to.equal( ImageStyleEditing.defaultIcons.right );
-			} );
-
-			it( 'should use one of default translations if #title matches', () => {
-				expect( plugin.imageStyles[ 2 ].title ).to.deep.equal( 'Default translation' );
-			} );
-
-			it( 'should extend one of default styles if #name matches', () => {
-				expect( plugin.imageStyles[ 3 ] ).to.deep.equal( {
-					name: 'imageStyleFull',
-					title: 'Custom title',
-					icon: ImageStyleEditing.defaultIcons.left,
-					isDefault: true
-				} );
-			} );
-		} );
-
-		describe( 'string format', () => {
-			it( 'should use one of default styles if #name matches', () => {
-				return VirtualTestEditor
-					.create( {
-						plugins: [ ImageStyleEditing ],
-						image: {
-							styles: [ 'imageStyleFull' ]
-						}
-					} )
-					.then( newEditor => {
-						editor = newEditor;
-						plugin = editor.plugins.get( ImageStyleEditing );
-						expect( plugin.imageStyles[ 0 ] ).to.deep.equal( ImageStyleEditing.defaultStyles.imageStyleFull );
-					} );
-			} );
-
-			it( 'should warn if a #name not found in default styles', () => {
-				testUtils.sinon.stub( log, 'warn' );
-
-				return VirtualTestEditor
-					.create( {
-						plugins: [ ImageStyleEditing ],
-						image: {
-							styles: [ 'foo' ]
-						}
-					} )
-					.then( newEditor => {
-						editor = newEditor;
-						plugin = editor.plugins.get( ImageStyleEditing );
-
-						expect( plugin.imageStyles[ 0 ] ).to.deep.equal( {
-							name: 'foo'
-						} );
-
-						sinon.assert.calledOnce( log.warn );
-						sinon.assert.calledWithExactly( log.warn,
-							sinon.match( /^image-style-not-found/ ),
-							{ name: 'foo' }
-						);
-					} );
-			} );
-		} );
-	} );
-
-	describe( 'localizedDefaultStylesTitles()', () => {
-		it( 'should return localized titles of default styles', () => {
-			return VirtualTestEditor
-				.create( {
-					plugins: [ ImageStyleEditing ]
-				} )
-				.then( newEditor => {
-					editor = newEditor;
-					plugin = editor.plugins.get( ImageStyleEditing );
-
-					expect( plugin.localizedDefaultStylesTitles ).to.deep.equal( {
-						'Full size image': 'Full size image',
-						'Side image': 'Side image',
-						'Left aligned image': 'Left aligned image',
-						'Centered image': 'Centered image',
-						'Right aligned image': 'Right aligned image'
-					} );
-				} );
-		} );
-	} );
-
-	describe( 'defaultStyles', () => {
-		it( 'should be defined', () => {
-			expect( ImageStyleEditing.defaultStyles ).to.deep.equal( {
-				imageStyleFull: {
-					name: 'imageStyleFull',
-					title: 'Full size image',
-					icon: fullWidthIcon,
-					isDefault: true
-				},
-				imageStyleSide: {
-					name: 'imageStyleSide',
-					title: 'Side image',
-					icon: rightIcon,
-					className: 'image-style-side'
-				},
-				imageStyleAlignLeft: {
-					name: 'imageStyleAlignLeft',
-					title: 'Left aligned image',
-					icon: leftIcon,
-					className: 'image-style-align-left'
-				},
-				imageStyleAlignCenter: {
-					name: 'imageStyleAlignCenter',
-					title: 'Centered image',
-					icon: centerIcon,
-					className: 'image-style-align-center'
-				},
-				imageStyleAlignRight: {
-					name: 'imageStyleAlignRight',
-					title: 'Right aligned image',
-					icon: rightIcon,
-					className: 'image-style-align-right'
-				}
-			} );
-		} );
-	} );
-
-	describe( 'defaultIcons', () => {
-		it( 'should be defined', () => {
-			expect( ImageStyleEditing.defaultIcons ).to.deep.equal( {
-				full: fullWidthIcon,
-				left: leftIcon,
-				right: rightIcon,
-				center: centerIcon,
-			} );
 		} );
 	} );
 } );
