+++ tests/imagestyle/imagestyleui.js
@@ -8,9 +8,12 @@ import ImageStyleEditing from '../../src/imagestyle/imagestyleediting';
 import ImageStyleUI from '../../src/imagestyle/imagestyleui';
 import ButtonView from '@ckeditor/ckeditor5-ui/src/button/buttonview';
 import global from '@ckeditor/ckeditor5-utils/src/dom/global';
+import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
+import VirtualTestEditor from '@ckeditor/ckeditor5-core/tests/_utils/virtualtesteditor';
 
 describe( 'ImageStyleUI', () => {
 	let editor;
+
 	const styles = [
 		{ name: 'style 1', title: 'Style 1 title', icon: 'style1-icon', isDefault: true },
 		{ name: 'style 2', title: 'Style 2 title', icon: 'style2-icon', cssClass: 'style2-class' },
@@ -60,6 +63,35 @@ describe( 'ImageStyleUI', () => {
 		}
 	} );
 
+	it( 'should translate buttons if taken from default styles', () => {
+		const editorElement = global.document.createElement( 'div' );
+
+		global.document.body.appendChild( editorElement );
+
+		class TranslationMock extends Plugin {
+			init() {
+				sinon.stub( this.editor, 't' ).returns( 'Default title' );
+			}
+		}
+
+		return ClassicTestEditor
+			.create( editorElement, {
+				plugins: [ TranslationMock, ImageStyleEditing, ImageStyleUI ],
+				image: {
+					styles: [
+						{ name: 'style 1', title: 'Side image', icon: 'style1-icon', isDefault: true }
+					]
+				}
+			} )
+			.then( newEditor => {
+				editor = newEditor;
+
+				const buttonView = editor.ui.componentFactory.create( 'style 1' );
+
+				expect( buttonView.label ).to.equal( 'Default title' );
+			} );
+	} );
+
 	it( 'should not add buttons to image toolbar if configuration is present', () => {
 		const editorElement = global.document.createElement( 'div' );
 		global.document.body.appendChild( editorElement );
@@ -73,8 +105,30 @@ describe( 'ImageStyleUI', () => {
 				}
 			} )
 			.then( newEditor => {
-				expect( newEditor.config.get( 'image.toolbar' ) ).to.eql( [ 'foo', 'bar' ] );
+				expect( newEditor.config.get( 'image.toolbar' ) ).to.deep.equal( [ 'foo', 'bar' ] );
 				newEditor.destroy();
 			} );
+	} );
+
+	describe( 'localizedDefaultStylesTitles()', () => {
+		it( 'should return localized titles of default styles', () => {
+			return VirtualTestEditor
+				.create( {
+					plugins: [ ImageStyleUI ]
+				} )
+				.then( newEditor => {
+					editor = newEditor;
+
+					const plugin = editor.plugins.get( ImageStyleUI );
+
+					expect( plugin.localizedDefaultStylesTitles ).to.deep.equal( {
+						'Full size image': 'Full size image',
+						'Side image': 'Side image',
+						'Left aligned image': 'Left aligned image',
+						'Centered image': 'Centered image',
+						'Right aligned image': 'Right aligned image'
+					} );
+				} );
+		} );
 	} );
 } );
